+++ src/Constants.ts
@@ -14,12 +14,11 @@ export const HIDE_CODE_TIME_STATUS_LABEL = 'Hide Code Time status';
 export const SHOW_CODE_TIME_STATUS_LABEL = 'Show Code Time status';
 
 // Directory where software data is stored
-export const SOFTWARE_DIRECTORY = '.software';
-export const api_endpoint = 'https://api.software.com';
-export const app_url = 'https://app.software.com';
+export const SOFTWARE_DIRECTORY = '.software-local'; // '.software';
+export const api_endpoint = 'http://localhost:5000';//'https://api.software.com';
+export const app_url = 'http://localhost:3000';//'https://app.software.com';
 
 export const vscode_issues_url = 'https://github.com/swdotcom/swdc-vscode/issues';
 
 export const TWENTY_SEC_TIMEOUT_MILLIS = 20000;
 export const ONE_MIN_MILLIS = 1000 * 60;
-export const FIVE_SECONDS_IN_MILLIS = 1000 * 5;
