+++ src/websockets.ts
@@ -7,7 +7,6 @@ import { handleFlowStateMessage } from './message_handlers/flow_state';
 import { userDeletedCompletionHandler } from './DataController';
 import { setEndOfDayNotification } from './notifications/endOfDay';
 import { handleAuthenticatedPluginUser } from './message_handlers/authenticated_plugin_user';
-import { getAuth0Instance } from './auth/Auth0AuthenticationProvider';
 
 const WebSocket = require('ws');
 
@@ -225,8 +224,10 @@ const handleIncomingMessage = (data: any) => {
         }
         break;
       case 'authenticated_plugin_user':
-        if (!getAuth0Instance()) {
-          handleAuthenticatedPluginUser(message.body);
+        const user = message.body;
+        const currentEmail = getItem('name');
+        if (user.email !== currentEmail) {
+          handleAuthenticatedPluginUser(user);
         }
         break;
       case 'flow_state':
